chore: refresh project audit guidance
